refactor(js-algo-ds/linked-list): reformat singly linked list
